+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * ⚠ **Breaking Change**: Changed ResultCode into an enum. Please adjust your usage of the PaymentResult (PaymentSessionFinished or PaymentAdvancedFinished). 
 * Updated iOS SDK to v5.7.1.
 * Updated Android SDK to v5.3.1.
+* Added support for customizing Drop-in and card component.
+* Added option to adjust Drop-in preselected payment method title.
 
 ## 0.0.3